Extracted a package. Give a group flag to the authorized keys file
