+++ user-authorization-builder/go.mod
@@ -4,12 +4,7 @@ go 1.13
 
 require (
 	github.com/go-ldap/ldap/v3 v3.2.2
-	github.com/golang/protobuf v1.4.2 // indirect
-	github.com/imdario/mergo v0.3.9 // indirect
-	golang.org/x/oauth2 v0.0.0-20200107190931-bf48bf16ab8d // indirect
-	golang.org/x/time v0.0.0-20200630173020-3af7569d3a1e // indirect
+	github.com/google/go-cmp v0.4.0 // indirect
 	k8s.io/api v0.18.5
 	k8s.io/apimachinery v0.18.5
-	k8s.io/client-go v11.0.0+incompatible
-	k8s.io/utils v0.0.0-20200619165400-6e3d28b6ed19 // indirect
 )
